+++ app/src/main/java/com/app/arcabyolimpo/data/remote/api/ArcaApi.kt
@@ -4,28 +4,35 @@ import com.app.arcabyolimpo.data.remote.dto.auth.LoginRequestDto
 import com.app.arcabyolimpo.data.remote.dto.auth.LoginResponseDto
 import com.app.arcabyolimpo.data.remote.dto.auth.RefreshRequestDto
 import com.app.arcabyolimpo.data.remote.dto.auth.RefreshResponseDto
-import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
-import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordDto
 import com.app.arcabyolimpo.data.remote.dto.beneficiaries.BeneficiariesListDto
 import com.app.arcabyolimpo.data.remote.dto.beneficiaries.BeneficiaryDto
-import com.app.arcabyolimpo.data.remote.dto.supplies.SuppliesListDto
-import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyDto
+import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
+import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordDto
 import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordDto
 import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.VerifyTokenResponseDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.GetFiltersDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.WorkshopCategoryListDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.SuppliesListDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyBatchDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.SupplyDto
 import com.app.arcabyolimpo.data.remote.dto.user.UserDto
 import com.app.arcabyolimpo.data.remote.dto.user.registeruser.RegisterResponseDto
 import com.app.arcabyolimpo.data.remote.dto.user.registeruser.RegisterUserDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.AddNewWorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopsListDto
+import com.app.arcabyolimpo.domain.model.supplies.SupplyBatchExt
+import okhttp3.MultipartBody
+import okhttp3.RequestBody
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.DELETE
 import retrofit2.http.GET
+import retrofit2.http.Multipart
 import retrofit2.http.POST
+import retrofit2.http.Part
 import retrofit2.http.Path
 import retrofit2.http.Query
 
@@ -63,17 +70,17 @@ interface ArcaApi {
 
     @POST("user/recover-password")
     suspend fun recoverPassword(
-        @Body request: RecoverPasswordDto
+        @Body request: RecoverPasswordDto,
     ): Response<RecoverPasswordResponseDto>
 
     @GET("user/verify-token")
     suspend fun verifyToken(
-        @Query("token") token: String
+        @Query("token") token: String,
     ): Response<VerifyTokenResponseDto>
 
     @POST("user/update-password")
     suspend fun updatePassword(
-        @Body request: UpdatePasswordDto
+        @Body request: UpdatePasswordDto,
     ): Response<UpdatePasswordResponseDto>
 
 
@@ -98,20 +105,44 @@ interface ArcaApi {
     suspend fun getWorkshopsList(): List<WorkshopsListDto>
 
     @GET("workshop/{id}")
-    suspend fun getWorkshop(@Path("id") id: String): WorkshopDto
+    suspend fun getWorkshop(
+        @Path("id") id: String,
+    ): WorkshopDto
 
     @POST("workshop/add")
     suspend fun addWorkshop(
-        @Body requestBody: WorkshopDto
+        @Body requestBody: WorkshopDto,
     ): AddNewWorkshopDto
 
     @GET("beneficiary")
     suspend fun getBeneficiariesList(): List<BeneficiariesListDto>
 
     @GET("beneficiary/{id}")
-    suspend fun getBeneficiary(@Path("id") id: String): BeneficiaryDto
+    suspend fun getBeneficiary(
+        @Path("id") id: String,
+    ): BeneficiaryDto
 
     @DELETE("beneficiary/{id}")
-    suspend fun deleteBeneficiary(@Path("id") id: String): BeneficiaryDto
+    suspend fun deleteBeneficiary(
+        @Path("id") id: String,
+    ): Response<Unit>
 
+    @GET("supplyBatch/{id}")
+    suspend fun getSupplyBatchById(
+        @Path("id") id: String,
+    ): SupplyBatchDto
+
+    @GET("supplies/workshop/category")
+    suspend fun getWorkshopCategoryList(): WorkshopCategoryListDto
+
+    @Multipart
+    @POST("supplies/add")
+    suspend fun addSupply(
+        @Part("idWorkshop") idWorkshop: RequestBody,
+        @Part("name") name: RequestBody,
+        @Part("measureUnit") measureUnit: RequestBody,
+        @Part("idCategory") idCategory: RequestBody,
+        @Part("status") status: RequestBody,
+        @Part image: MultipartBody.Part?
+    )
 }