+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/user/UserDto.kt
@@ -6,30 +6,17 @@ data class UserDto(
     @SerializedName("idUsuario") val idUsuario: String? = null,
 
     @SerializedName("idRol") val idRol: String,
-
     @SerializedName("nombre") val nombre: String,
-
     @SerializedName("apellidoPaterno") val apellidoPaterno: String,
-
     @SerializedName("apellidoMaterno") val apellidoMaterno: String,
-
     @SerializedName("fechaNacimiento") val fechaNacimiento: String,
-
     @SerializedName("carrera") val carrera: String,
-
     @SerializedName("correoElectronico") val correoElectronico: String,
-
     @SerializedName("contrasena") val contrasena: String? = null,
-
     @SerializedName("telefono") val telefono: String,
-
     @SerializedName("estatus") val estatus: Int,
-
     @SerializedName("reglamentoInterno") val reglamentoInterno: Int? = null,
-
     @SerializedName("copiaINE") val copiaINE: Int? = null,
-
     @SerializedName("avisoConfidencialidad") val avisoConfidencialidad: Int? = null,
-
-    @SerializedName("foto") val foto: String? = null
+    @SerializedName("foto") val foto: String? = null,
 )