+++ app/src/main/java/com/app/arcabyolimpo/data/repository/user/UsersRepositoryImpl.kt
@@ -7,6 +7,8 @@ import com.app.arcabyolimpo.data.remote.dto.user.UserDto
 import com.app.arcabyolimpo.domain.repository.user.UsersRepository
 import javax.inject.Inject
 
+// Repository implementation that interacts with the remote API to fetch user data.
+// Converts the received DTOs into domain models using the user mapper before returning them.
 
 class UsersRepositoryImpl
     @Inject
@@ -15,14 +17,14 @@ class UsersRepositoryImpl
     ) : UsersRepository {
         override suspend fun getUsers(): List<UserDto> {
             val response = api.getAllUsers()
-            return response.map{ it.toDomain() }
+            return response.map { it.toDomain() }
         }
 
     override suspend fun getUserById(id: String): UserDto {
-        return try {
+         try {
             val response = api.getUserById(id)
             if (response.isNotEmpty()) {
-                response.first().toDomain()
+                return response.first().toDomain()
             } else {
                 throw Exception("Usuario no encontrado")
             }