+++ app/src/main/java/com/app/arcabyolimpo/domain/repository/supplies/SupplyRepository.kt
@@ -1,8 +1,12 @@
 package com.app.arcabyolimpo.domain.repository.supplies
 
+import android.net.Uri
 import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
 import com.app.arcabyolimpo.domain.model.filter.FilterData
 import com.app.arcabyolimpo.domain.model.supplies.Supply
+import com.app.arcabyolimpo.domain.model.supplies.SupplyAdd
+import com.app.arcabyolimpo.domain.model.supplies.SupplyBatchExt
+import com.app.arcabyolimpo.domain.model.supplies.WorkshopCategoryList
 
 /**
  * Retrieves detailed information for a specific supply by its [id].
@@ -16,7 +20,16 @@ interface SupplyRepository {
 
     suspend fun getSupplyById(id: String): Supply
 
+    suspend fun getSupplyBatchById(id: String): SupplyBatchExt
+
     suspend fun filterSupply(params: FilterDto): List<Supply>
 
     suspend fun getFilterData(): FilterData
+
+    suspend fun getWorkshopCategoryList(): Result<WorkshopCategoryList>
+
+    suspend fun addSupply(
+        supply: SupplyAdd,
+        image: Uri?
+    ): Result<Unit>
 }