+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/passwordregisteration/PasswordRegistrationScreen.kt
@@ -30,7 +30,21 @@ import com.app.arcabyolimpo.presentation.ui.components.atoms.inputs.StandardInpu
 import com.app.arcabyolimpo.ui.theme.Background
 import com.app.arcabyolimpo.ui.theme.White
 
-
+/**
+ * Composable screen for registering a new password to activate a user account.
+ *
+ * This screen allows users to create and confirm a new password after receiving
+ * an account activation link or recovery email. It includes input validation,
+ * error handling, and feedback on success.
+ *
+ * The screen observes state updates from [PasswordRegistrationViewModel] and navigates
+ * to the success destination when the password registration process completes successfully.
+ *
+ * @param email The user's email address used for password registration.
+ * @param onBackClick Callback triggered when the user taps the back navigation button.
+ * @param onPasswordRegistrationSucessClick Callback invoked when password registration succeeds.
+ * @param viewModel The [PasswordRegistrationViewModel] providing state and business logic.
+ */
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Suppress("ktlint:standard:function-naming")
@@ -62,14 +76,14 @@ fun PasswordRegistrationScreen(
                         ReturnIcon(size = 16.dp)
                     }
                 },
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = Background,
-                ),
+                colors =
+                    TopAppBarDefaults.topAppBarColors(
+                        containerColor = Background,
+                    ),
             )
         },
-        containerColor = Background
-    )
-    {   padding ->
+        containerColor = Background,
+    ) { padding ->
         Column(
             modifier =
                 Modifier
@@ -80,15 +94,16 @@ fun PasswordRegistrationScreen(
             Text(
                 text = "El Arca en Querétaro I.A.P",
                 style = Typography.bodyMedium,
-                color = White
+                color = White,
             )
             Spacer(modifier = Modifier.height(20.dp))
             Text(
-                text = "Crea una \n" +
+                text =
+                    "Crea una \n" +
                         "contraseña para \n" +
                         "activar tu cuenta",
                 style = Typography.headlineLarge,
-                color = White
+                color = White,
             )
             Spacer(modifier = Modifier.height(80.dp))
             StandardInput(
@@ -142,9 +157,8 @@ fun PasswordRegistrationScreen(
     }
 }
 
-//@Preview(showBackground = true, showSystemUi = true)
-//@Composable
-//fun PasswordRegistrationScreenPreview() {
+// @Preview(showBackground = true, showSystemUi = true)
+// @Composable
+// fun PasswordRegistrationScreenPreview() {
 //    PasswordRegistrationScreen(onBackClick = { })
-//}
-
+// }