+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/passwordregisteration/PasswordRegistrationSuccessScreen.kt
@@ -22,28 +22,37 @@ import com.app.arcabyolimpo.ui.theme.Background
 import com.app.arcabyolimpo.ui.theme.ErrorRed
 import com.app.arcabyolimpo.ui.theme.White
 
+/**
+ * Composable screen displayed when the user successfully registers a new password.
+ *
+ * This screen confirms that the password setup process has been completed and encourages
+ * the user to log in with their new credentials. It maintains a consistent visual design
+ * with the rest of the authentication flow using the app’s theme and typography.
+ *
+ * @param onBackClick Callback triggered when the user taps the back navigation button.
+ */
+
+@Suppress("ktlint:standard:function-naming")
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun PasswordRegistrationSuccessScreen (
-    onBackClick: () -> Unit,
-) {
+fun PasswordRegistrationSuccessScreen(onBackClick: () -> Unit) {
     Scaffold(
-    topBar = {
-        TopAppBar(
-            title = {},
-            navigationIcon = {
-                IconButton(onClick = onBackClick) {
-                    ReturnIcon(size = 16.dp)
-                }
-            },
-            colors = TopAppBarDefaults.topAppBarColors(
-                containerColor = Background,
-            ),
-        )
-    },
-    containerColor = Background
-    )
-    {   padding ->
+        topBar = {
+            TopAppBar(
+                title = {},
+                navigationIcon = {
+                    IconButton(onClick = onBackClick) {
+                        ReturnIcon(size = 16.dp)
+                    }
+                },
+                colors =
+                    TopAppBarDefaults.topAppBarColors(
+                        containerColor = Background,
+                    ),
+            )
+        },
+        containerColor = Background,
+    ) { padding ->
         Column(
             modifier =
                 Modifier
@@ -54,15 +63,16 @@ fun PasswordRegistrationSuccessScreen (
             Text(
                 text = "El Arca en Querétaro I.A.P",
                 style = Typography.bodyMedium,
-                color = White
+                color = White,
             )
             Spacer(modifier = Modifier.height(20.dp))
             Text(
-                text = "¡Listo! \n" +
+                text =
+                    "¡Listo! \n" +
                         "Intenta\n" +
                         "iniciar sesión",
                 style = Typography.headlineLarge,
-                color = White
+                color = White,
             )
         }
     }